requirements.txt
asgiref==3.8.1
Django==3.2.13
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
gunicorn==20.1.0
django-cors-headers==3.7.0